requirements.txt
asgiref==3.2.10
Django==3.1.2
djangorestframework==3.12.1
joblib==0.17.0
mysqlclient @ file:///C:/Users/Anish/OneDrive/Desktop/workon_project/BRS/mysqlclient-1.4.6-cp37-cp37m-win32.whl
numpy==1.19.2
pandas==1.1.3
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
